requirements.txt
# LightRAG 및 관련 의존성
lightrag-hku>=1.3.8
nano-vectordb>=0.0.4.3
numpy>=1.24.0

# LLM 서비스
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Ollama 지원
ollama>=0.1.7

# 문서 처리
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
markdown>=3.5.1
chardet>=5.2.0

# 유틸리티
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
typer>=0.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 한국어 처리
kiwipiepy>=0.17.1
konlpy>=0.6.0

# 로깅
loguru>=0.7.2

# 비동기 처리
aiofiles>=23.2.1

# 테스트
pytest>=8.0.0
pytest-asyncio>=0.23.5 